feat(table): 更新表测试中的整数类型支持
- 在导入语句中添加了Int64类型支持
- 将硬编码的整数值100替换为Int64(100)构造函数调用
- 确保表操作中的数据类型一致性
- 提高了测试代码的数据类型安全性

+++ table_list_test.py
@@ -7,7 +7,7 @@ from free_list import FreeList, new_free_list_from_page_id, new_free_list
 from row import new_row
 from table_list import Table, TableSeqGenerator, new_table_seq_generator, TableList, new_table_list_from_page_id, \
     new_table_list, new_table
-from utils import from_buf
+from utils import Int64, from_buf
 from value.const import VALUE_TYPE_STRING, VALUE_TYPE_INT
 from value.int import new_int
 from value.string import new_string
@@ -74,7 +74,7 @@ def test_table():
     name = inspect.currentframe().f_code.co_name
     fd, table = init_table(name)
     key = new_int(10)
-    row = new_row(100, [new_string("xiaoming"), new_string("m"), new_int(90)])
+    row = new_row(Int64(100), [new_string("xiaoming"), new_string("m"), new_int(90)])
     table.set(key, row)
     row = table.get(key)
     row.show()
